+++ src/components/Calendar/interfaces/interfaces.ts
@@ -19,11 +19,13 @@ export interface IDatePickerProps extends IDatePickerBaseProps {
   setDate?: Date | null;
   startDate?: Date | null;
   endDate?: Date | null;
+  isSelectTime?: boolean;
 }
 
 export interface IDatePickerChooserProps extends IDatePickerBaseProps {
   value: string;
   isActive: boolean;
+  isTimeSelect?: boolean;
 }
 
 export interface IRangeForDatePickerState {